quell braindead compiler warning (vc)
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@7255 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/Player.cxx
@@ -595,7 +595,7 @@ void			Player::setDeadReckoning(float timestamp)
   } else if (deadReckoningState == 1) {
     alpha     = 0.5;
   } else {
-    alpha     = 0.1;
+    alpha     = (float)0.1;
   }
   // alpha filtering
   deltaTime = deltaTime + offset * alpha;
